refactor: move download manager out of core
