+++ src/core/index.ts
@@ -20,18 +20,6 @@ export type { ScriptOptions, GeneratedScript } from './packager/script-generator
 
 export { FileSplitter, getFileSplitter } from './packager/file-splitter';
 
-// Download Manager
-export { DownloadManager, getDownloadManager } from './download-manager';
-export type {
-  DownloadManagerItem,
-  DownloadManagerItemStatus,
-  DownloadManagerOptions,
-  DownloadManagerResult,
-  OverallProgress,
-  DownloadManagerEvents,
-} from './download-manager';
-export type { DownloadItem, DownloadItemStatus, DownloadOptions, DownloadResult } from './download-manager';
-
 // Resolvers
 export { PipResolver, getPipResolver } from './resolver/pip-resolver';
 export { MavenResolver, getMavenResolver } from './resolver/maven-resolver';
